Implement greedy variant: done
Testing has shown that the greedy variant often misses words that
even the normal decoding recognizes with high probability.

This is because once a word is detected as active, even if a better
suited start comes along it will be ignored leading to misalignment.
This is even more an issue because audio token borders are fuzzy:
the start of a word may lie 'between' tokens so token idx 8 and 9
might share the same phonemes.

We need to improve this somehow!